requirements.txt
Django>=1.4
mimeparse>=0.1.3
python-dateutil==1
django-tastypie==0.9.11
pystache==0.3.1
